chatbot-single-agent-multiple-tools/BE/requirements.txt
fastapi
pydantic
uvicorn
openai
pymilvus
sentence-transformers
python-dotenv
pandas